ovpn: prevent sleep in RCU locked area
strp_data_ready() calls ovpn_tcp_rcv() which calls ovpn_peer_del()
in case of TCP errors.
The latter invokes ovpn_socket_release() which may sleep.

This causes a sleep in atomic context.

Fix this misbehaviour by taking the reference to the peer
outside of the RCU locked area, so that sleeping is allowed
again.

Reported-by: Qingfang Deng <dqfext@gmail.com>
Closes: https://github.com/OpenVPN/ovpn-net-next/issues/13
Signed-off-by: Antonio Quartulli <antonio@openvpn.net>